+++ source/Application/Controller/ArticleListController.php
@@ -674,7 +674,7 @@ class ArticleListController extends \OxidEsales\Eshop\Application\Controller\Fro
         // assign template name
         if (($templateName = basename(Registry::getRequest()->getRequestEscapedParameter('tpl')))) {
             $this->_sThisTemplate = 'custom/' . $templateName;
-        } elseif (($category = $this->getActiveCategory()) && $category->oxcategories__oxtemplate->value) {
+        } elseif (($category = $this->getActiveCategory()) && $category->getFieldData('oxtemplate')) {
             $this->_sThisTemplate = $category->oxcategories__oxtemplate->value;
         }