+++ source/Application/Model/NewsSubscribed.php
@@ -167,7 +167,7 @@ class NewsSubscribed extends \OxidEsales\Eshop\Core\Model\BaseModel
      */
     public function getOptInStatus()
     {
-        return $this->oxnewssubscribed__oxdboptin->value;
+        return (int) $this->getFieldData('oxdboptin');
     }
 
     /**